+++ src/files/Import.js
@@ -2,12 +2,16 @@ import ImportOBJ from './ImportOBJ.js';
 import ImportSGL from './ImportSGL.js';
 import ImportPLY from './ImportPLY.js';
 import ImportSTL from './ImportSTL.js';
+import ImportGLTF from './ImportGLTF.js';
 
 var Import = {
   importOBJ: ImportOBJ.importOBJ,
   importSGL: ImportSGL.importSGL,
   importPLY: ImportPLY.importPLY,
-  importSTL: ImportSTL.importSTL
+  importSTL: ImportSTL.importSTL,
+  importGLTF: ImportGLTF.importGLTF,
+  importGLB: ImportGLTF.importGLB
 };
 
 export default Import;
+